update to render()
In the render function, render attempts would go through _async_render() and, if and only if content was generated, eventually close the new page created. However, if the attempt timed out, the page would not be closed, yet a new attempt would open a new page, leaving behind an extra page. This is my first proposal ever on github, sorry if it's messy.